+++ plugins/moxie/src/index.ts
@@ -178,8 +178,5 @@ export function createTenoxUI(config: CreatorConfig = {}) {
 
 export * from './types'
 export * from './utils'
-export * from './lib/regexp'
-export { transform } from './lib/transformer'
-export { Processor } from './lib/processor'
-export { Renderer } from './lib/renderer'
+export * from './lib'
 export default Moxie